Implement signed integers
This commit adds support for arbitrary-width signed integers.

As discussed in #14 `Int::new()` and `Int::value()` both work with values
that may exceed the specified bit width (if they are negative).
Equivelant methods for working with bitfields have been added
(`Int::to_bits` and `Int::from_bits`), together with documentation
explaining this behaviour.

Unfortunately this results in a fair bit of code duplication,
there does not seem to be a good way to abstract over the differences
between signed and unsigned numbers while keeping the API const-friendly.

Stuff missing for feature-parity with unsigned integers:
* Saturating arithmetic operations.
* Checked arithmetic operations.
* Overflowing arithmetic operations.
* Byte operations (`to_ne_bytes`, `extract_iN`, etc).
* Support for the various optional crate features.