Add `serde` support for signed integers (#56)
* Add `serde` support to `Int`

The serialized value comes from `Int::value()` rather than
`Int::to_bits()`. This means the value may exceed the type's bit-width.

* Avoid storing `UInt::MAX` in `InvalidUIntValueError`

This reduces the size of `InvalidUIntValueError` in-memory by a factor of
two, which could slightly improve performance in some scenarios.

As the bounds on the `Serialize`/`Deserialize` implementations remain
equivelant from the users point of view, this is not a breaking change.