Implement `extract_i{8,16,32,64,128}` functions for signed integers
This is implemented as a common macro to reduce the amount of code
 repetition. `UInt` also makes use of said macro, as it generates
functionally identical code to what was previously written out by hand.

Having these functions be generated by a macro also makes it a lot less
work to introduce e.g. `try_extract_i<N>` methods in the future.