Fix borsh serialization/deserialization with "slow" writers/readers
`borsh::io::Read`/`borsh::io::Write` can read/write fewer bytes than
requested, which leaves the input/output buffer only partially filled.
This happens for example if a `TcpStream` is waiting on the other end to
send more data. Instead of blocking, `Read` will return after writing
just the data that is currently available.

The implementation of `BorshDeserialize`/`BorshSerialize` previously
didn't take this into account: it'd assume the entire buffer got filled
after one call. This commit fixes that and adds a test case for it.

It also asserts (at compile time) that the intermediate byte buffer is
large enough to fit the `UInt`. This (currently) cannot fail for any of
the types provided by this crate, but checking seems better than silently
truncating the number.