Unconditionally declare the crate as `#[no_std]`
By unconditionally declaring this crate as `no_std` we opt out of the
standard library's prelude, which implicitly brings items like `Vec` and
`String` into scope.

Since we'd need to import those manually in case the `alloc` crate is
used but the standard library is disabled, we might as well keep things
consistent and always manually import them.

Note that the standard library API's are still accessible when the `std`
feature is enabled. This is currently not used anywhere, but the option
is there in case STL-specific functionality is introduced in the future.