Add various extra tests to CI (#74)
* Fix clippy errors and run rustfmt

The `since` field of the `Number` deprecation attribute wasn't a valid
semantic version, which result in a `clippy::deprecated_semver` error.

* Test for clippy warnings in CI

* Test that all Rust code is properly formatted in CI

It might be worth also checking `Cargo.toml` with something like Taplo,
but that can be done at a later date.

* Test for rustdoc warnings in CI

Rustdoc has a few lints that aren't included in `cargo check`.

* Test for semantic versioning violations in CI with cargo-semver-checks