Implement borsh for signed integers (#77)
* Implement borsh for signed integers

This is pretty similar to https://github.com/danlehmann/arbitrary-int/pull/69 but
takes advantage of Integer traits a little more. In particular to make this happen,
two new functions from_unsigned and to_unsigned are now available to the Trait,
which allow turning a SignedInteger into an UnsignedInteger (with the same number
of bits).

* Address review